Rename Spinner and FitBadge components for consistency
Renamed spinner.tsx to SpinnerElement.tsx and fitbadge.tsx to Badge.tsx for improved naming consistency. Updated all imports and the central index export to reflect these changes.
